add file and fix

+++ discount-service/src/main/java/com/service/discount/requests/DiscountRequest.java
@@ -19,7 +19,7 @@ public class DiscountRequest {
   private UUID id;
   private String discountCode;
 
-  private Long discountType;
+  private Integer discountType;
 
   private String discountTitle;
   private BigDecimal discountPercentage;
