+++ discount-service/src/main/java/com/service/discount/requests/TargetRequest.java
@@ -2,6 +2,8 @@ package com.service.discount.requests;
 
 import java.util.UUID;
 
+import org.bouncycastle.util.Integers;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -14,7 +16,7 @@ import lombok.Setter;
 @Setter
 @Builder
 public class TargetRequest {
-  private Long id;
+  private Integer id;
   private UUID discountId;
   private String targetType;
   private UUID targetId;
